# OCaml API: revise upper bounds checks which use pow