--- a/bindings/ocaml/EverCrypt.ml
+++ b/bindings/ocaml/EverCrypt.ml
@@ -138,7 +138,7 @@ module Hash = struct
   open EverCrypt_Hash
   module Noalloc = struct
     let hash ~alg ~msg ~digest =
-      check_max_input_len alg (C.size msg);
+      check_max_buffer_len (C.size msg);
       assert (C.size digest = digest_len alg);
       assert (C.disjoint digest msg);
       everCrypt_Hash_hash (alg_definition alg) (C.ctypes_buf digest) (C.ctypes_buf msg) (C.size_uint32 msg)
@@ -155,10 +155,10 @@ module Hash = struct
     let incr_len = ref Z.zero in
     Gc.finalise everCrypt_Hash_Incremental_free st;
     (alg, incr_len, st)
-  let update ~st:(alg, incr_len, t) ~msg =
-    check_max_input_len alg (C.size msg);
+  let update ~st:(_alg, incr_len, t) ~msg =
+    check_max_buffer_len (C.size msg);
     incr_len := Z.add !incr_len (Z.of_int (C.size msg));
-    assert (Z.lt !incr_len (max_input_len alg));
+    assert (Z.lt !incr_len max_uint64);
     everCrypt_Hash_Incremental_update t (C.ctypes_buf msg) (C.size_uint32 msg)
   let finish ~st:(alg, incr_len, t) =
     let digest = C.make (digest_len alg) in
@@ -191,8 +191,8 @@ module HMAC = struct
       (* Hacl.HMAC.compute_st *)
       assert (C.size tag = HashDefs.digest_len alg);
       assert (C.disjoint msg tag);
-      HashDefs.check_key_len alg (C.size key);
-      HashDefs.check_key_len alg (C.size msg);
+      check_max_buffer_len (C.size key);
+      check_max_buffer_len (C.size msg);
       everCrypt_HMAC_compute (HashDefs.alg_definition alg) (C.ctypes_buf tag) (C.ctypes_buf key) (C.size_uint32 key) (C.ctypes_buf msg) (C.size_uint32 msg)
   end
   let mac ~alg ~key ~msg =
@@ -233,16 +233,16 @@ module HKDF = struct
       assert (C.size prk = HashDefs.digest_len alg);
       assert (C.disjoint salt prk);
       assert (C.disjoint ikm prk);
-      HashDefs.check_key_len alg (C.size salt);
-      HashDefs.check_key_len alg (C.size ikm);
+      check_max_buffer_len (C.size salt);
+      check_max_buffer_len (C.size ikm);
       everCrypt_HKDF_extract (HashDefs.alg_definition alg) (C.ctypes_buf prk) (C.ctypes_buf salt) (C.size_uint32 salt) (C.ctypes_buf ikm) (C.size_uint32 ikm)
     let expand ~alg ~prk ~info ~okm =
       (* Hacl.HKDF.expand_st *)
       assert (C.size okm <= 255 * HashDefs.digest_len alg);
       assert (C.disjoint okm prk);
       assert (HashDefs.digest_len alg <= C.size prk);
-      HashDefs.(check_max_input_len alg (digest_len alg + block_len alg + C.size info + 1));
-      HashDefs.check_key_len alg (C.size prk);
+      check_max_buffer_len (C.size info);
+      check_max_buffer_len (C.size prk);
       everCrypt_HKDF_expand (HashDefs.alg_definition alg) (C.ctypes_buf okm) (C.ctypes_buf prk) (C.size_uint32 prk) (C.ctypes_buf info) (C.size_uint32 info) (C.size_uint32 okm)
   end
   let extract ~alg ~salt ~ikm =