While profiling the memory usage of a piece of code which uses the incremental EverCrypt hash interface, @polubelova noticed that the runtime checks in the OCaml API, specifically those comparing the size of a buffer to a large power of 2, result in excessive allocations.

~~The original intention was to encode the F* preconditions as runtime checks, but I think in this case the cost outweighs the benefits, especially since the large upper bound being tested here are unlikely to be reached in many real-world use cases. I therefore propose to remove them.~~

As suggested in this PR, I've hoisted computations of the relevant powers of 2 and replaced `Z.pow` with `Z.shift_left`. I've also reduced the checked bounds to be no larger than the types that represent them. This solves the excessive allocations problem.