@franziskuskiefer 2022-05-31
I have to look at the actual code. But shifting could indeed be used to check the length without allocating memory. It would be nice if we could keep the checks.