@msprotz 2022-05-31
@polubelova defensive APIs in C is another matter, and potentially a fairly involved piece of work (but I agree it'd be nice to have).

Regarding the specific issue of a slow left-shift: there are multiple things you can do here without disabling safety checks.
- Stop recomputing `pow2 64` on every call; can you define `let pow2_64 = pow2 64` at the top-level and use that?
- More efficient `pow2` using `Z.shift_left`

If none of these work according to your benchmarks, then we can maybe contemplate having an "insane" mode where some checks are disabled but not by default to keep the API safe in all cases.

Can you investigate a little bit more this pow2 issue? Unfortunately I don't have access to you benchmarks so I don't know how to reproduce. Thanks!