@victor-dumitrescu 2022-05-31
Makes sense, I can try rewriting the checks with `shift_left`. Marina also pointed out that some of these checks aren't doing anything in practice (see comment above) and could be rewritten with lower bounds than in the specs.