@victor-dumitrescu 2022-09-15
I was running some benchmarks and thought it might be useful to document the improvement that we got from this fix, specifically for Blake2b. Lower memory usage overall + for very small digests it makes a noticeable running time difference too:

- before:
```
│ Hacl/Blake2b_256.hash/0              │     433.39ns │  37.00w │ 0.14e-3 │
│ Hacl/Blake2b_256.hash/100            │     439.87ns │  37.00w │ 0.14e-3 │
│ Hacl/Blake2b_256.hash/1000           │   1_651.76ns │  37.00w │ 0.14e-3 │
│ Hacl/Blake2b_256.hash/10000          │  13_904.17ns │  37.00w │ 0.14e-3 │
│ Hacl/Blake2b_256.hash/100000         │ 136_611.97ns │  37.00w │ 0.14e-3 │
│ Hacl/Blake2b_32.hash/0               │     597.80ns │  37.00w │ 0.14e-3 │
│ Hacl/Blake2b_32.hash/100             │     597.11ns │  37.00w │ 0.14e-3 │
│ Hacl/Blake2b_32.hash/1000            │   2_903.03ns │  37.00w │ 0.14e-3 │
│ Hacl/Blake2b_32.hash/10000           │  25_962.63ns │  37.00w │ 0.14e-3 │
│ Hacl/Blake2b_32.hash/100000          │ 252_820.96ns │  37.00w │ 0.13e-3 │
```

- after:
```
│ Hacl/Blake2b_256.hash/0              │     303.42ns │  31.00w │ 0.12e-3 │
│ Hacl/Blake2b_256.hash/100            │     307.70ns │  31.00w │ 0.12e-3 │
│ Hacl/Blake2b_256.hash/1000           │   1_565.54ns │  31.00w │ 0.12e-3 │
│ Hacl/Blake2b_256.hash/10000          │  14_128.73ns │  31.00w │ 0.12e-3 │
│ Hacl/Blake2b_256.hash/100000         │ 138_708.93ns │  31.00w │ 0.12e-3 │
│ Hacl/Blake2b_32.hash/0               │     461.02ns │  31.00w │ 0.12e-3 │
│ Hacl/Blake2b_32.hash/100             │     465.45ns │  31.00w │ 0.12e-3 │
│ Hacl/Blake2b_32.hash/1000            │   2_747.43ns │  31.00w │ 0.12e-3 │
│ Hacl/Blake2b_32.hash/10000           │  25_470.37ns │  31.00w │ 0.12e-3 │
│ Hacl/Blake2b_32.hash/100000          │ 250_708.27ns │  31.00w │ 0.10e-3 │
```